Método para exceções

+++ src/main/java/com/devari/loginUsuario/service/UsuarioService.java
@@ -13,24 +13,32 @@ public class UsuarioService {
 
     private final UsuarioRepository usuarioRepository;
 
-    public List<Usuario> findByAll() {
+    public List<Usuario> findAll() {
         return usuarioRepository.findAll();
     }
 
     public Usuario save(Usuario usuario) {
-        if (usuario.getName() == null || usuario.getName().isBlank()){
-            throw new IllegalArgumentException("name is empty or null");
-        }
+        verifyEmptyOrNull(usuario);
         return usuarioRepository.save(usuario);
     }
 
     public void deleteById(Long id) {
+        if (!usuarioRepository.existsById(id)){
+            throw new IllegalArgumentException("usuario id not exist");
+        }
         usuarioRepository.deleteById(id);
     }
 
     public Usuario upload(Long id, Usuario usuario){
+        verifyEmptyOrNull(usuario);
         Usuario usuarioDb = usuarioRepository.findById(id).orElseThrow(() -> new RuntimeException("Usuario not found"));
         usuarioDb.setName(usuario.getName());
         return usuarioRepository.save(usuarioDb);
+    }
+
+    public void verifyEmptyOrNull(Usuario usuario){
+        if (usuario.getName() == null || usuario.getName().isBlank()){
+            throw new IllegalArgumentException("name is empty or null");
+        }
     }
 }
